+++ nessi/signal/__init__.py
@@ -6,6 +6,7 @@ from __future__ import (absolute_import,
 from .windowing import time_window
 from .windowing import space_window
 from .tapering import taper1d
+from .filtering import sin2filter
 
 if __name__ == '__main__':
     import doctest